+++ src/lib/RbtLigandSiteMapper.cxx
@@ -142,7 +142,8 @@ RbtCavityList RbtLigandSiteMapper::operator()() {
   }
 
   // Find the contiguous regions of cavity grid points
-  RbtFFTPeakMap peakMap = spGrid->FindPeaks(cavVal, minSize);
+  RbtFFTPeakMap peakMap =
+      spGrid->FindPeaks(cavVal, static_cast<unsigned int>(minSize));
 
   // Convert peaks to cavity format
   for (RbtFFTPeakMapConstIter pIter = peakMap.begin(); pIter != peakMap.end();
